### 2024-02-19 #6193 @sneakers-the-rat
@lsandig I see the main repository is at sourcehut - https://sr.ht/~lsandig/Kirstine.jl/

typically a review will involve raising issues/PRs in the repository, and that serves both as a record of the review and as a way to coordinate work with the authors. I want to respect your choice of git host, but also satisfy the archival needs of JOSS which is currently on github. Do you have preferences or guides for us here - would raising the issues on github be ok for the purposes of review, or is there some way of mirroring them between sourcehut and github?

@arfon is there a way we usually handle this?
