### 2024-02-19 #6193 @arfon
@sneakers-the-rat – our only conditions about git hosting are that it's possible for people to open issues/provide feedback, and that there isn't a manual process for account creation (e.g., institutional hosts).

As @lsandig has created a mirror, I think it's OK for the review feedback to be provided on the GitHub mirror (https://github.com/lsandig/Kirstine.jl) but at the end of the review we might want to update the URL of the repository for the paper back to the sourcehut URL (https://sr.ht/~lsandig/Kirstine.jl/).
