+++ paper/paper.bib
@@ -7,8 +7,8 @@
   number =	 1,
   pages =	 {65--98},
   year =	 2017,
-  url =		 {https://doi.org/10.1137/141000671},
   publisher =	 {SIAM},
+  doi =          {10.1137/141000671},
 }
 
 @manual{bornkamp-2023-dosef,
@@ -48,6 +48,7 @@
   journal =	 {Computer Methods and Programs in Biomedicine},
   volume =	 74,
   year =	 2004,
+  doi =          {10.1016/s0169-2607(03)00073-7},
 }
 
 @manual{masoudi-2020-icaod,
@@ -69,6 +70,7 @@
   volume =	 108,
   year =	 2012,
   url =		 {https://CRAN.R-project.org/package=PopED},
+  doi =          {10.1016/j.cmpb.2012.05.005},
 }
 
 @article{overstall-2020-acebayes,
