### 2024-02-19 #6193 @sneakers-the-rat
Fabulous. and it looks like there is at least one way to mirror issues back to sourcehut if you want to keep issues there for your own records. https://github.com/aThorp96/sourcehut_issue_mirror

I'll start searching for reviewers!
