### 2024-02-19 #6193 @sneakers-the-rat
I wonder if @jflournoy , @coryc3133 , or @dcosme (hi!) as experimental design/modeling experts speak julia and would be interested in reviewing this?

Otherwise I see @lbluque and @Dom-Owens-UoB listing themselves as knowing Julia and regressino models.
