### 2024-02-20 #6193 @Coryc3133
Sorry, I also don't speak Julia (yet), unfortunately. Sounds like
interesting work!

On Tue, Feb 20, 2024, 1:01 PM John Flournoy ***@***.***>
wrote:

> Very intriguing work. I don't speak Julia, though. Not yet.
>
