### 2024-02-26 #6193 @harisorgn
Hello, this seems very interesting, I would be happy to review!

I've never reviewed for JOSS before, so would appreciate any guidance in the process. I'm comfortable in Julia and Bayesian statistics though, so that's something 😅 .
