### 2024-02-27 #6193 @sneakers-the-rat
excellent!!!! I will give you more information once we find a second reviewer and get started, and of course will be available throughout the process for questions. For now the review guides are here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
here: https://joss.readthedocs.io/en/latest/review_criteria.html
and here! https://joss.readthedocs.io/en/latest/review_checklist.html

@lsandig do you have any other potential reviewers to recommend?
