### 2024-02-27 #6193 @harisorgn
I could ask around the Julia language Slack if that's useful? There are dedicated channels to statistics & Bayesian methods.
