### 2024-02-28 #6424 @editorialbot
[REVIEW]: Kirstine.jl: A Julia Package for Bayesian Optimal Design of Experiments
**Submitting author:** @lsandig (<a href="http://orcid.org/0000-0002-3174-3275">Ludger Sandig</a>)
**Repository:** https://github.com/lsandig/Kirstine.jl
**Branch with paper.md** (empty if default branch): paper-draft
**Version:** v0.5.0
