### 2024-02-28 #6424 @sneakers-the-rat
OK welcome @harisorgn and @roualdes and thank you for your time here.

As the top issue says, you'll be working from a checklist, a [set of guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), and some [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html).

The goal here is to collaboratively help this package reach a basic standard for usability, maintainability, and correctness - we are all friends here, there are no gates to be kept, we're just trying to learn from and help each other out. Feel free to talk amongst each other as peers, this is an open and iterative review, so you don't need to wait until you've reviewed anything before raising an issue or commenting.

The checklist is a set of minimum standards that the package has to meet, but you are also free to read, comment, raise issues and pull requests, and do whatever else falls under your purview as a reviewer. Think about what you would want to have if you were a potential new user of the package - is the documentation clear? do the tests cover what they need to? etc. That kind of neutral third-party perspective is invaluable for making software accessible and maintainable.

As you review, i would suggest that you raise issues in the repository and link back to this issue so that we can track issues related to this review. Since github issues don't have threading, this is how we can keep discussion organized, because commenting in this issue can be a bit messy. Don't be shy commenting here to ask questions about the review itself, though! It's up to you all and the maintainers if you want to follow any sort of naming convention for issues/PRs. This repository is hosted on sourcehut and mirrored to github, but for the purposes of recordkeeping for this review, lets keep the issues on GH and we'll change the repo link to sourcehut at the end.

Time spent reviewing can range from a few hours just completing the checklist to longer depending on how much attention to detail you want to give - I have completed a review in as little as 1 hour when it was obvious everything was shipshape and the authors didn't need any help, and as long as several months when i was helping the authors establish docs and tests. That's up to you! Your time is your gift to give.

Other than that, i'll be here so feel free to ask questions! The first thing you'll want to do is generate your checklists using the commands in the OP comment
