### 2024-03-18 lsandig/Kirstine.jl#2 @roualdes
JOSS Review: feedback on paper
Over all the writing in this draft is nice.  I found the mathematical background easy to follow, and even prefer the write up here to the similarly named section of your doc.  The Statement of Need is clear and organized well.

Below are just a few comments, suggestions, and/or questions.

* Line 13 should change "computing" to "compute"
* Should references to $\Sigma$ include an argument $x$?
* Line 27 references an un-named prior distribution.  Later you reference what I believe is this same distribution, naming its density function $p(\theta)$.  Would it be better to name it on line 27?
* You say, "Having obtained a candidate design $𝜉^∗$, we then apply an equivalence theorem from infinite-dimensional convex analysis to verify that the design $\xi^*$ is indeed optimal."  Are you suggesting Kirstine.jl does this for you?  If so, I missed this part and think a push in the right direction would benefit readers.
* Are there general rules, guidelines, or strategies for guessing $K > K^*$ that could be cited?
* Starting on line 47, you say, "These matrix operations are a performance bottleneck, since functions in R are passed by value."  While I do agree with the overall idea here, it seems this sentence is too strong.  I believe R follows [copy-on-modify](https://adv-r.hadley.nz/names-values.html#copy-on-modify). Doesn't Kirstine also make new arrays after transforming $\Sigma(x)$ in each Monte Carlo summand?  I couldn't verify one way or the other from your code base.  Please do point me in the right direction.
