### 2024-03-18 #6424 @roualdes
I made it through the draft paper and filed [Issue #2](https://github.com/lsandig/Kirstine.jl/issues/2) with a few comments.
