### 2024-03-21 lsandig/Kirstine.jl#2 @lsandig
Thank you for the helpful feedback!

>     * Line 13 should change "computing" to "compute"

Fixed.

>     * Should references to Σ include an argument x?

That depends on what problem variant you look at. Compared to the full math writeup in the package docs, I did not want to introduce too much notation in the paper. Hence I'm already only describing the special case where T is the identity transformation, and where Σ is constant. I'm also glossing over the separation of covariate set X and design region Z.

I think I'll add a note that these generalizations are also possible with Kirstine.jl.

>     * Line 27 references an un-named prior distribution.  Later you reference what I believe is this same distribution, naming its density function p(θ).  Would it be better to name it on line 27?

Yes. I have reworded this part, and also indicated that p is a density on Θ.

>     * You say, "Having obtained a candidate design 𝜉∗, we then apply an equivalence theorem from infinite-dimensional convex analysis to verify that the design ξ∗ is indeed optimal."  Are you suggesting Kirstine.jl does this for you?  If so, I missed this part and think a push in the right direction would benefit readers.

Yes, this part of Kirstine.jl. Basically, the equivalence theorem is an infinite-dimensional variant of “ξ* is optimal ⇔ all directional derivatives at ξ* are nonpositive”, only that there are now infinitely many (linearly independent) directions. Kirstine.jl has a function for computing the derivative on a grid of directions, and you can then check their maximum. Typically, people also look at plots for this, which are implemented here: https://git.sr.ht/~lsandig/Kirstine.jl/tree/main/item/src/plot.jl#L215

I'll mention this feature in the paper.

>     * Are there general rules, guidelines, or strategies for guessing K>K∗ that could be cited?

For problems of locally optimal design there is an upper bound that only depends on the dimension of Θ. Unfortunately, to my knowledge, for Bayesian design problems there is no such bound.

>     * Starting on line 47, you say, "These matrix operations are a performance bottleneck, since functions in R are passed by value."  While I do agree with the overall idea here, it seems this sentence is too strong.  I believe R follows [copy-on-modify](https://adv-r.hadley.nz/names-values.html#copy-on-modify).

Thank you for the link, this part of the R internals was new to me. I'll think about a way to reword this.

> Doesn't Kirstine also make new arrays after transforming Σ(x) in each Monte Carlo summand?

No, it actually first updates, and then Cholesky-decomposes Σ in place:

https://git.sr.ht/~lsandig/Kirstine.jl/tree/main/item/src/model-nonlinear-regression.jl#L32

> I couldn't verify one way or the other from your code base.  Please do point me in the right direction.

To see the places where the pre-allocated matrices are used you can start in `objective!()`: https://git.sr.ht/~lsandig/Kirstine.jl/tree/main/item/src/designproblem.jl#L192

There, the following steps happen in place, without creating new arrays
- map design point z to covariate x
- compute Cholesky decomposition of Σ(x)
- for each θ in the sample:
  - compute D_θμ(x, θ)
  - compute M(ξ,θ)
  - potentially transform M
  - evaluate Φ
