### 2024-03-21 lsandig/Kirstine.jl#2 @lsandig
Here are the changes to the draft: https://github.com/lsandig/Kirstine.jl/commit/52e0fd1676eacab4ab145b047707c028f508a4d9
