### 2024-03-22 #6424 @roualdes
While checking Functionality > Performance, I compared Kirstine.jl to the R package [ICAOD](https://cran.r-project.org/web/packages/ICAOD/index.html), which references the associated paper [A Metaheuristic Adaptive Cubature Based Algorithm to Find Bayesian Optimal Designs for Nonlinear Models](https://doi.org/10.1080%2F10618600.2019.1601097).  In Section 4.2 Sigmoid Emax Model of the linked paper, with code in the supplementary materials, the authors use nearly the same model as in the Getting Started section of Kirstine.jl's doc.  The main differences, as far as I could tell, were the priors and the underlying algorithm used to find the optimal design.  Here's some comments relative to this experience.

* I could not force the priors in the R code to match the priors in Kirstine.jl's Getting Started example.  Instead, I could relative easily make Kisrtine.jl code match the examples in the R code.  This is a compliment to the documentation and design of Kirstine.jl
* The R code I ran was just slightly faster than my Julia code.  Because the two packages are using different underlying algorithms to find optimal designs, I don't think this conclusion is very meaningful.  And I agree with @lsandig's overall statements about Kirstine's design, namely the added ease/flexibility in Kirstine.jl's one language solution to these problems, unlike in R which has to resort to C/C++ for efficient code.
