### 2024-03-22 #6424 @roualdes
[Issue #3](https://github.com/lsandig/Kirstine.jl/issues/3) has some comments/questions related to doc.
