### 2024-03-26 lsandig/Kirstine.jl#3 @lsandig
> So the macros `@simple_model` and `@simple_parameter` appear to take as a first argument an arbitrary name and then append Model and Parameter, respectively, to the user specified first argument.  Is this correct? This isn't explicitly mentioned, but it's my best understanding of what is going.  Whether or not I'm right, it seems like there could be an extra sentence to help the user through this.

This behavior is documented [in the docstrings](https://lsandig.srht.site/Kirstine.jl/api.html#Helpers) of the two macros. I'll point this out more clearly when they are first used in the tutorial.

> Can `@simple_parameter` reference only scalars? Does Kirstine.jl allow blocks within a parameter vector to be associated with one named parameter?

It only takes scalars, since I think that a (mathematical) vector of real numbers is the most common use case, and here the dimension of the parameter space is simply the number of scalar arguments. For anything more complicated (e.g. matrix or simplex parameters, parameter types with extra flags to indicate elements that should be considered known, ...) I think it's clearer to specify the type and `dimension()` method manually.

> I found the two uses of M, namely M and MT, difficult to parse, especially once I got to the Gateaux Derivative section.

I'll try and see if I can find a better notation then; however, I've already exhausted most of the Latin and Greek alphabets. At least, using semantic math LaTeX macros for everything now pays off :)

> The JOSS draft paper and the doc Section Mathematical Background took reverse approaches to introduce complex equations. […]

Thanks for pointing this out. Restructuring the math page from a large reference info dump to something more readable from top to bottom is already on my TODO list.

> In the sub-section Rationale, you say "These are different tasks than what Kirstine.jl attempts to do."  I think a sentence or two in the second paragraph about what explicitly separates the use-case of Kirstine.jl from the use cases of the R packages found on CRAN could be helpful.

Ok, I'll make it clearer that these design packages are for ANOVA/linear/polynomial regression models, while Kirstine.jl deals with nonlinear models.
