### 2024-03-27 lsandig/Kirstine.jl#3 @lsandig
I've worked on the documentation for a bit, and incorporated your suggestions. The changes are on the [development branch](https://git.sr.ht/~lsandig/Kirstine.jl/log/develop) over at sourcehut, starting at 00aef6bb. I don't have rendered html documentation for the development branch set up, I hope this is not too inconvenient for you to read.
