### 2024-03-28 lsandig/Kirstine.jl#3 @roualdes
Great to see that you used some of my suggestions to update the doc.  I'm happy for you to close the issue at your convenience.
