### 2024-03-28 #6424 @roualdes
No blockers, as far as I can see.  The package, doc, and paper look of good quality and I'm happy to sign off on moving forward towards publication.

@sneakers-the-rat, what is the preferred way for a reviewer to signal that they have completed their review and everything looks good?
