### 2024-04-03 lsandig/Kirstine.jl#4 @harisorgn
JOSS Review 2 
Overall I found `Kirstine.jl` a great contribution to experimental design and to the Julia ecosystem. The motivation is clear and convincing, the codebase well organised and thoroughly tested and the documentation is extensive.

I am covered by past comments in #2  and #3 and am happy to sign off this article for publication. Only have a couple of typos for @lsandig consideration :

- minor typo, `1` and `2` superscripts should be flipped here to make the explanation after the equation less confusing :
https://github.com/lsandig/Kirstine.jl/blob/1e23cd0d1c98c3ba209b0a41a7b5acde4b3e3251/docs/src/math.md?plain=1#L428

- another (more) minor typo, should be `t_{\DimUnit}` instead of `1_...` here :
https://github.com/lsandig/Kirstine.jl/blob/1e23cd0d1c98c3ba209b0a41a7b5acde4b3e3251/docs/src/dtr.md?plain=1#L62
