### 2024-04-03 #6424 @harisorgn
Apologies for the delay, I have finished going over the article, code and documentation. I took several notes along the way and then deleted them as @lsandig had already addressed them elsewhere. In particular I appreciate the well tested & organised codebase and the thorough documentation, which I am hoping should reduce friction even for non-experienced Julia users (now with the accompanying `kirstine-examples` repository)!

In the end I only had typo-based comments in [Issue #4](https://github.com/lsandig/Kirstine.jl/issues/4). Happy to sign off for publication too.
