### 2024-04-03 lsandig/Kirstine.jl#4 @harisorgn
>So, I kept the notation and definition, but tried to make the interpretation a bit clearer: https://git.sr.ht/~lsandig/Kirstine.jl/commit/d5b556b

You are right, I confused myself, sorry 😅 . Thanks for the change.
