### 2024-05-06 #6424 @sneakers-the-rat
Alright @lsandig see the above post for final tasks - need to make a new tagged release and archive it on zenodo or elsewhere, i'll get started on the rest of the tasks
