### 2024-05-06 #6424 @sneakers-the-rat
Confirmed that DoseFinding and ICAOD don't have archival versions with DOIs (and DoseFinding is written by the author, so presumably they would know the correct citation).

I don't see any typos in manuscript and trust reviewers to have evaluated it for content.

@lsandig let me know when you have deposited the archival version and i'll complete my portion of the review and recommend the paper for acceptance :)
