### 2024-05-13 #6424 @sneakers-the-rat
standby @lsandig, we'll make this work with the sourcehut repo <3
