### 2024-05-13 #6424 @sneakers-the-rat
OK @lsandig having a problem where the git repo from sourcehut works, but it doesn't seem to have the `paper-draft` branch that the github repo does. i assumed that the github repo was a 1:1 mirror of the sourcehut repository, but could you add the paper to the sourcehut repo (in another branch is still fine) and then let me know when you do?
